+++ R/get_model_stats.R
@@ -156,6 +156,109 @@ get_model_stats_splm <- function(cov_est_object, data_object, estmethod) {
   )
 }
 
+get_model_stats_splm_iid <- function(cov_est_object, data_object, estmethod) {
+
+
+  X <- do.call("rbind", data_object$X_list)
+  y <- do.call("rbind", data_object$y_list)
+  qr_val <- qr(X)
+  R_val <- qr.R(qr_val)
+  s2 <- cov_est_object$spcov_params_val[["ie"]]
+  cor_betahat <- chol2inv(chol(crossprod(R_val, R_val)))
+  cov_betahat <- s2 * cor_betahat
+  betahat <- as.numeric(backsolve(R_val, qr.qty(qr_val, y)))
+  names(betahat) <- colnames(data_object$X_list[[1]])
+  cov_betahat <- as.matrix(cov_betahat)
+  rownames(cov_betahat) <- colnames(data_object$X_list[[1]])
+  colnames(cov_betahat) <- colnames(data_object$X_list[[1]])
+  fitted <- X %*% betahat
+  resids <- y - fitted
+
+
+  # return coefficients
+  coefficients <- get_coefficients(betahat, cov_est_object$spcov_params_val, cov_est_object$randcov_params_val)
+
+  # return fitted
+  fitted <- list(
+    response = as.numeric(fitted),
+    spcov = list(de = as.numeric(rep(0, length(y))), ie = as.numeric(resids)),
+    randcov = NULL
+  )
+
+
+  # return hat values
+  hatvalues <- diag(X %*% tcrossprod(cor_betahat, X))
+  # return residuals
+  residuals <- list(
+    response = as.numeric(resids),
+    pearson = 1 / sqrt(s2) * resids
+  )
+  residuals$standardized <- residuals$pearson / sqrt(1 - hatvalues)
+
+  # return cooks distance
+  cooks_distance <- get_cooks_distance(residuals, hatvalues, data_object$p)
+
+  # reorder relevant quantities
+  ## fitted
+  fitted$response <- fitted$response[order(data_object$order)]
+  names(fitted$response) <- data_object$observed_index
+  fitted$spcov$de <- fitted$spcov$de[order(data_object$order)]
+  names(fitted$spcov$de) <- data_object$observed_index
+  fitted$spcov$ie <- fitted$spcov$ie[order(data_object$order)]
+  names(fitted$spcov$ie) <- data_object$observed_index
+  hatvalues <- hatvalues[order(data_object$order)]
+  names(hatvalues) <- data_object$observed_index
+  residuals$response <- residuals$response[order(data_object$order)]
+  names(residuals$response) <- data_object$observed_index
+  residuals$pearson <- residuals$pearson[order(data_object$order)]
+  names(residuals$pearson) <- data_object$observed_index
+  residuals$standardized <- residuals$standardized[order(data_object$order)]
+  names(residuals$standardized) <- data_object$observed_index
+  cooks_distance <- cooks_distance[order(data_object$order)]
+  names(cooks_distance) <- data_object$observed_index
+
+  # return variance covariance matrices
+  vcov <- get_vcov(cov_betahat)
+
+  # return deviance
+  deviance <- as.numeric(crossprod(residuals$pearson, residuals$pearson))
+  muhat <- mean(y)
+  pearson_null <- 1 / sqrt(s2) * (y - muhat)
+  deviance_null <- as.numeric(crossprod(pearson_null, pearson_null))
+  pseudoR2 <- as.numeric(1 - deviance / deviance_null)
+
+  # set null model R2 equal to zero (no covariates)
+  if (length(labels(terms(data_object$formula))) == 0) {
+    pseudoR2 <- 0
+  }
+
+
+  ## empirical semivariogram
+  if (estmethod == "sv-wls") {
+    esv_val <- cov_est_object$esv
+  } else {
+    esv_val <- NULL
+  }
+
+
+  # npar
+  p_theta_spcov <- length(cov_est_object$is_known$spcov) - sum(cov_est_object$is_known$spcov)
+  p_theta_randcov <- length(cov_est_object$is_known$randcov) - sum(cov_est_object$is_known$randcov)
+  npar <- p_theta_spcov + p_theta_randcov
+
+  list(
+    coefficients = coefficients,
+    fitted = fitted,
+    hatvalues = hatvalues,
+    residuals = residuals,
+    cooks_distance = cooks_distance,
+    vcov = vcov,
+    deviance = deviance,
+    pseudoR2 = pseudoR2,
+    npar = npar
+  )
+}
+
 
 get_model_stats_spautor <- function(cov_est_object, data_object, estmethod) {
   # cov_est_object$randcov_params_val is NULL if not added so won't affect downstream calculations